feat(oa4rust): complete all core module stubs — CMS, BAM, process platform, file, BBS, org control

+++ oa4rust/crates/control/src/person.rs
@@ -109,7 +109,7 @@ async fn query_page(
             .query(
                 "SELECT id, unique_id, name, mobile, email, locked FROM auth_person \
                  WHERE deleted_at IS NULL AND (unique_id > $1 OR $1 = '' OR $1 = '-') \
-                 ORDER BY unique_id ASC LIMIT $2",
+                 ORDER BY unique_id ASC LIMIT $2::bigint",
                 &[&flag, &limit],
             )
             .await
@@ -121,7 +121,7 @@ async fn query_page(
             .query(
                 "SELECT id, unique_id, name, mobile, email, locked FROM auth_person \
                  WHERE deleted_at IS NULL AND (unique_id < $1 OR $1 = '' OR $1 = '-') \
-                 ORDER BY unique_id DESC LIMIT $2",
+                 ORDER BY unique_id DESC LIMIT $2::bigint",
                 &[&flag, &limit],
             )
             .await
